pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rtbench"
version = "0.1.0"
description = "RepoRT retention-time transfer benchmark"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "joblib>=1.3,<2",
  "lightgbm>=4.0,<5",
  "numpy>=1.24,<3",
  "pandas>=2.0,<3",
  "PyYAML>=6.0,<7",
  "scikit-learn>=1.3,<2",
  "scipy>=1.10,<2",
  "torch>=2.0,<3",
  "xgboost>=1.7,<3",
]

[project.optional-dependencies]
test = [
  "coverage>=7.13",
  "pytest>=8.3",
  "pytest-cov>=7.1",
]

[project.scripts]
rtbench-run = "rtbench.run:main"
rtbench-merge-runs = "rtbench.merge_runs:main"
rtbench-experiments = "rtbench.experiments:main"
rtbench-report-vs-unirt = "rtbench.report_vs_unirt:main"

[tool.setuptools.packages.find]
include = ["rtbench*"]

[tool.pytest.ini_options]
minversion = "8.3"
testpaths = ["tests"]
markers = [
  "smoke: end-to-end smoke tests that run the real benchmark pipeline",
]
addopts = "-ra"
